+++ patzilla/util/network/browser.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 # (c) 2017-2019 Andreas Motl <andreas.motl@ip-tools.org>
 
-regular_user_agent = u'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_3) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/68.0.3421.0 Safari/537.36'
+regular_user_agent = u'Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:100.0) Gecko/20100101 Firefox/100.0'