+++ CHANGES.rst
@@ -34,6 +34,7 @@ Development
 - [mw] Honour ``fulltext_enabled`` setting of data source
 - [mw] Disable acquisition of PDF documents from DPMA/DEPATISconnect
 - [mw] EPO/OPS: Unlock inquiry of image information for documents without kind code
+- [mw] EPO/Espacenet: Repair access layer
 
 
 2019-11-01 0.169.3